requirements.txt
autopep8==1.7.0
click==8.1.3
colorama==0.4.5
Flask==2.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pycodestyle==2.9.1
pyparsing==3.0.9
pytesseract==0.3.10
toml==0.10.2
Werkzeug==2.2.2
zipp==3.8.1
